implement test_period_helper
